we are only going to support in kernel query for Michael work, no resource to support more
