+++ include/Makefile
@@ -1,7 +1,3 @@
 install:
 	cp --force ./camquery.h /usr/include/camquery.h
-	cp --force ./camquery_edge.h /usr/include/camquery_edge.h
 	cp --force ./camquery_kernel.h /usr/include/camquery_kernel.h
-	cp --force ./camquery_node.h /usr/include/camquery_node.h
-	cp --force ./camquery_service.h /usr/include/camquery_service.h
-	cp --force ./camquery_utils.h /usr/include/camquery_utils.h
